WifiHAL: Optimized by packing multiple log entries into an event
In the old design, an event contains only a log entry, it is not
efficient for indicating logs. As a simple idea, it can be optimized
that packing multiple log entries into an event.

To support the compatibility of old/new mechanisms, check the magic
number of packet header to determine which approach will be applied.

Bug: 285446447
Test: Build Pass and verify done
Defer-CP-To-Master: 285446447
Change-Id: I1481f10d2da3f29bb03941e445961474d605db26